КОНТРОЛЛЕР / Отдельная ф-ия для создания пользователя
